requirements.txt
asgiref==3.8.1
Brotli==1.1.0
dj-database-url==2.3.0
Django==5.2.1
django-admin-sortable2==2.2.8
pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
